+++ vault/secret.go
@@ -109,7 +109,6 @@ func (secret *vaultSecret) Renew(client *vault.Client, leaseIdSuffix string) (Me
 	rawSecret, err := client.Sys().Renew(leaseId, 0)
 	if err != nil {
 		log.Printf("the secret with lease ID %s could not be renewed", leaseId)
-		log.Print(err)
 		return Metadata{}, err
 	}
 	log.Printf("the lease for %s was successfully renewed", leaseId)