+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.2.4 (Next)
+### 1.2.4
 - Minor code optimization.
 - Explicitly marshal Concourse metadata values as Strings.
 - Fix error returned for KV secret with nonexistent specified version.